clients/python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
beautifulsoup4==4.12.3
    # via bs4
black==24.8.0
    # via -r requirements.in
bs4==0.0.2
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   elastic-transport
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via black
elastic-transport==8.15.0
    # via elasticsearch
elasticsearch==8.15.1
    # via -r requirements.in
exceptiongroup==1.2.2
    # via pytest
flake8==7.1.1
    # via -r requirements.in
idna==3.10
    # via requests
iniconfig==2.0.0
    # via pytest
mccabe==0.7.0
    # via flake8
mypy==1.11.2
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==24.1
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via black
pluggy==1.5.0
    # via pytest
pycodestyle==2.12.1
    # via flake8
pyflakes==3.2.0
    # via flake8
pytest==8.3.3
    # via
    #   -r requirements.in
    #   pytest-dependency
pytest-dependency==0.6.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
soupsieve==2.6
    # via beautifulsoup4
tomli==2.0.1
    # via
    #   black
    #   mypy
    #   pytest
types-requests==2.32.0.20240914
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   black
    #   mypy
urllib3==2.2.3
    # via
    #   elastic-transport
    #   requests
    #   types-requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
